improve interface, esp. auth

+++ cmccdb_interface/client/edit.py
@@ -41,33 +41,63 @@ import os
 import flask
 from ..database import manage, datasets, backups
 
-from . import handlers
+from . import handlers, authentication
 
 bp = flask.Blueprint("edit", __name__, url_prefix="/client", template_folder=".")
 
 WEB_UPLOADER_NAME = "Web Uploader"
 WEB_UPLOADER_EMAIL = "maboyer@tamu.edu"
+WEB_UPLOADER_ENV_NAME = "CMCCDB_UPLOADER_NAME"
+WEB_UPLOADER_ENV_EMAIL = "CMCCDB_UPLOADER_EMAIL"
+def get_uploader_info():
+    name = flask.request.args.get("name")
+    username = flask.request.args.get("username")
+    email = flask.request.args.get("email")
+
+    if name is None:
+        if username is not None and " " in username:
+            name = username
+        else:
+            name = os.environ.get(WEB_UPLOADER_ENV_NAME, WEB_UPLOADER_NAME)
+    if email is None:
+        name = os.environ.get(WEB_UPLOADER_ENV_EMAIL, WEB_UPLOADER_EMAIL)
+    
+    return {
+        "name":name,
+        "email":email
+    }    
+
+STAGING_DATABASE = "staging"
 @bp.route("/api/upload", methods=["POST"])
 def upload_dataset():
     """Writes the request body to the datasets table without validation."""
 
     database_name = flask.request.args.get("database")
-    default_uploader = os.environ.get("CMCCDB_UPLOADER_NAME", WEB_UPLOADER_NAME)
-    default_email = os.environ.get("CMCCDB_UPLOADER_EMAIL", WEB_UPLOADER_EMAIL)
-
     try:
+        auth_info = authentication.gh_get_cache_user_info()
+        if auth_info is None:
+            raise ValueError("unauthenticated users can't upload data")
+        if database_name != STAGING_DATABASE and not auth_info["member"]:
+            raise ValueError("unauthenticated users can't upload data")
+
+        uploader_info = get_uploader_info()
         file_name = flask.request.files['uploadFile'].filename
         body = flask.request.files['uploadFile'].read()
 
         dataset = datasets.prep_and_create_pb_dataset(
             file_name,
             body,
-            uploader_name=flask.request.args.get("user", default_uploader),
-            uploader_email=flask.request.args.get("email", default_email)
+            uploader_name=uploader_info["name"],
+            uploader_email=uploader_info["email"]
             )
         manage.add_dataset(dataset, database_name=database_name)
-        # backups.commit_backup()
-        return flask.jsonify({"redirect_url":"browse"})
+        try:
+            backups.git_backup()
+        except OSError as e:
+            from cmccdb_schema.logging_helpers import get_logger
+            logger = get_logger('uploads')
+            logger.exception("Backup failed, maybe")
+        return {'dataset_id':dataset.dataset_id}
     except Exception as error:  # pylint: disable=broad-except
         return flask.abort(handlers.make_error_response(error, 406))
 
@@ -77,7 +107,13 @@ def reconfigure_database():
 
     database_name = flask.request.args.get("database")
     try:
-        manage.reset_database(database_name, rebuild=True)
+        auth_info = authentication.gh_get_cache_user_info()
+        if auth_info is None:
+            raise ValueError("unauthenticated users can't reconfigure databases")
+        if not auth_info["owner"]:
+            raise ValueError("only CMCC admins can reconfigure databases")
+
+        manage.reset_database(database_name, force_quit=True)
         manage.configure_database(database_name)
         return handlers.make_string_response("ok")
     except Exception as error:  # pylint: disable=broad-except
@@ -87,6 +123,12 @@ def reconfigure_database():
 def delete_dataset(dataset_id):
     """Writes the request body to the datasets table without validation."""
     try:
+        auth_info = authentication.gh_get_cache_user_info()
+        if auth_info is None:
+            raise ValueError("unauthenticated users can't reconfigure databases")
+        if not auth_info["owner"]:
+            raise ValueError("only CMCC admins can reconfigure databases")
+
         database_name = flask.request.args.get("database")
         manage.delete_dataset(dataset_id, database_name=database_name)
         return handlers.make_string_response("ok")
@@ -102,15 +144,21 @@ def test_endpoint(endpoint, name):
         sys.path.insert(0, ENDPOINT_FOLDER)
     
     try:
+        # auth_info = authentication.gh_get_cache_user_info()
+        # if auth_info is None:
+        #     raise ValueError("unauthenticated users can't use the dev endpoints")
+        # if not auth_info["owner"]:
+        #     raise ValueError("only CMCC admins can use the dev endpoints databases")
+
         caller = None
         if endpoint+".py" in os.listdir(ENDPOINT_FOLDER):
             if endpoint in sys.modules:
                 mod = importlib.reload(sys.modules[endpoint])
             else:
                 mod = importlib.import_module(endpoint)
-            caller = getattr(mod, name)
+            caller = getattr(mod, name, None)
         if caller is None:
-            raise ImportError(f"No endpoint module {endpoint}")
+            raise ImportError(f"No endpoint {endpoint}.{name}")
         return caller()
 
     except Exception as error:  # pylint: disable=broad-except
